pyproject.toml
[tool.poetry]
name = "feishu-notifier-bot"
version = "0.1.0"
description = ""
authors = ["dxsooo <zeohan.dxs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
requests = "^2.27.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
